CDHSH-80 Fixed issue with sticky sidebar, tests and rebased
