+++ services/frontend/src/components/SkillsBadges/SkillBadges.scss
@@ -1,0 +1,19 @@
+@import "styles/_colors.scss";
+@import "styles/_dimens.scss";
+@import "styles/_mixins.scss";
+
+.skill-badge {
+    margin: $default-spacing / 2;
+}
+
+.text-badge {
+    background-color: $color-neutral-8;
+    border-radius: $default-border-radius;
+    color: $color-text-primary;
+    padding: $default-spacing/2 $default-spacing;
+
+    &.text-badge--highlighted {
+        background-color: $color-secondary;
+        color: $color-light;
+    }
+}
