+++ services/frontend/src/components/SkillsBadges/index.js
@@ -1,0 +1,1 @@
+export {default} from "./SkillBadges";
