+++ services/frontend/src/components/index.js
@@ -7,3 +7,4 @@ export {default as Navbar} from "./Navbar";
 export {default as NavSidebar} from "./NavSidebar";
 export {default as ProjectCard} from "./ProjectCard";
 export {default as ScrollContainer} from "./ScrollContainer";
+export {default as SkillBadges} from "./SkillsBadges";
